refactor: move some more code
